+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 0.2.0
 
-- New `AESystemSound.Path` enum
+- New enum with all system sound paths (iOS 11.4.1)
 
 ## Version 0.1.0